adds video player features

+++ controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from '../models/videoModel';
+import Comment from "../models/commentModel";
 import routes from '../routes';
 
 export const home = async (req, res) => {
@@ -15,7 +16,7 @@ export const search = async (req, res) => {
   const {term: searchingBy} = req.query;
   let videos = [];
   try {
-    videos = await Video.find({ title: {$regex: searchingBy, $options: 'i'} })
+    videos = await Video.find({title: {$regex: searchingBy, $options: 'i'}});
   } catch (error) {
     console.log(error);
   }
@@ -23,6 +24,7 @@ export const search = async (req, res) => {
 };
 
 export const getUpload = (req, res) => res.render('upload', {pageTitle: 'Upload'});
+
 export const postUpload = async (req, res) => {
   const {
     file: {path},
@@ -33,15 +35,18 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
   });
-  console.log(newVideo);
+
+  req.user.videos.push(newVideo._id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
 export const videoDetail = async (req, res) => {
   const {id} = req.params;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('creator').populate("comments");
     res.render('videoDetail', {pageTitle: video.title, video});
   } catch (error) {
     console.log(error);
@@ -53,7 +58,11 @@ export const getEditVideo = async (req, res) => {
   const {id} = req.params;
   try {
     const video = await Video.findById(id);
-    res.render('editVideo', {pageTitle: `Edit ${video.title}`, video});
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render('editVideo', {pageTitle: `Edit ${video.title}`, video});
+    }
   } catch (error) {
     console.log(error);
     res.redirect(routes.home);
@@ -78,9 +87,54 @@ export const postEditVideo = async (req, res) => {
 export const deleteVideo = async (req, res) => {
   const {id} = req.params;
   try {
-    await Video.findOneAndRemove({_id: id});
+    const video = await Video.findById(id);
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({_id: id});
+    }
   } catch (error) {
     console.log(error);
   }
   res.redirect(routes.home);
-}
+};
+
+// Register Video View
+export const postRegisterView = async (req, res) => {
+  const {
+    params: {id},
+  } = req;
+  try {
+    const video = await Video.findById(id);
+    video.views += 1;
+    video.save();
+    res.status(200);
+  } catch (error) {
+    res.status(400);
+  } finally {
+    res.end();
+  }
+};
+
+// Add Comment
+export const postAddComment = async (req, res) => {
+  const {
+    params: {id},
+    body: {comment},
+    user,
+  } = req;
+  try {
+    const video = await Video.findById(id);
+    const newComment = await Comment.create({
+      text: comment,
+      creator: user.id,
+    });
+    video.comments.push(newComment.id);
+    video.save();
+  } catch (error) {
+    console.log(error);
+    res.status(400);
+  } finally {
+    res.end();
+  }
+};
